// src/app/_shared/interfaces/candidate.ts
export interface Candidate {
    id: string;
    name: string;
    lastName: string;
    email: string;
    phone: string;
    isAvailable: boolean;
    city: string;
    technology: string;
    experience: Date;
    isRemote: boolean;
    comments: string;
    linkedIn: string;
    noticePeriod: number;
}